Реализация возможности редактирования
